refactor : git 충돌 해결
