+++ run/src/main/java/com/guide/run/global/security/config/SecurityConfig.java
@@ -43,7 +43,7 @@ public class SecurityConfig {
                     .requestMatchers("/member-upload")
                     .requestMatchers("/event-upload")
                     .requestMatchers("/attendance-upload")
-                    .requestMatchers("/api/oauth/**");
+                    .requestMatchers("/api/**");
         };
     }
     @Bean
